Authorization

+++ app/Http/Controllers/IdeaController.php
@@ -12,6 +12,11 @@ class IdeaController extends Controller
     }
     public function edit(Idea $idea)
     {
+        if(auth()->id() !== $idea->user_id)
+        {
+            abort(404);
+        }
+
         $editing = true;
 
         return view('Ideas.show', compact('idea', 'editing'));
@@ -26,7 +31,7 @@ class IdeaController extends Controller
         $validated = request()->validate([
             'content' =>'required|min:5|max:255',
         ]);
-
+        $validated['user_id'] = auth()->id();
         Idea::create($validated);
 
         return redirect()->route('dashboard')->with('success', 'Idea Created Successfully');
@@ -34,6 +39,11 @@ class IdeaController extends Controller
 
     public function destroy(Idea $idea)
     {
+        if(auth()->id() !== $idea->user_id)
+        {
+            abort(404);
+        }
+
         $idea->delete();
 
         return redirect()->route('dashboard')->with('success', 'Item Deleted Successfully');
@@ -41,6 +51,11 @@ class IdeaController extends Controller
 
     public function update(Idea $idea)
     {
+        if(auth()->id() !== $idea->user_id)
+        {
+            abort(404);
+        }
+        
         $validated = request()->validate([
             'content' =>'required|min:5|max:255',
         ]);
